--- packages/react/src/components/Textarea/useTextarea.ts
@@ -1,51 +1,0 @@
-import type { TextareaHTMLAttributes } from 'react';
-import { useContext } from 'react';
-
-import type { FormField } from '../../utilities/hooks/useFormField/useFormField';
-import { useFormField } from '../../utilities/hooks/useFormField/useFormField';
-import { FieldsetContext } from '../Fieldset/FieldsetContext';
-
-import type { TextareaProps } from './Textarea';
-
-type UseTextarea = (props: TextareaProps) => Omit<FormField, 'inputProps'> & {
-  textareaProps: Pick<
-    TextareaHTMLAttributes<HTMLTextAreaElement>,
-    'readOnly' | 'name' | 'required' | 'onClick' | 'onChange'
-  > &
-    FormField['inputProps'];
-};
-
-/** Handles props for `Textarea` in context with `Fieldset` */
-export const useTextarea: UseTextarea = (props) => {
-  const fieldset = useContext(FieldsetContext);
-  const {
-    inputProps,
-    readOnly,
-    size = fieldset?.size ?? 'md',
-    ...rest
-  } = useFormField(props, 'textarea');
-
-  return {
-    ...rest,
-    readOnly,
-    size,
-    textareaProps: {
-      ...inputProps,
-      readOnly,
-      onClick: (e) => {
-        if (readOnly) {
-          e.preventDefault();
-          return;
-        }
-        props?.onClick?.(e);
-      },
-      onChange: (e) => {
-        if (readOnly) {
-          e.preventDefault();
-          return;
-        }
-        props?.onChange?.(e);
-      },
-    },
-  };
-};
